Included library folder for easier sharing and testing. Fixed a few warnings about type conversions and unreferenced parameters.

+++ ParticleRewindData.h
@@ -6,7 +6,7 @@ struct ParticleRewindData
 	~ParticleRewindData();
 
 
-	int countAlive_{ 0 };
+	size_t countAlive_{ 0 };
 
 	uint color_{ 0 };
